requirements.txt
django
djangorestframework
django-cors-headers
elasticsearch-dsl
psycopg2-binary
celery
redis
